+++ xlang/gobuildserver/import_path.go
@@ -30,19 +30,22 @@ func (n *noGoGetDomainsT) reconfigure() {
 	n.mu.Lock()
 	defer n.mu.Unlock()
 
-	n.domains = strings.Split(conf.Get().NoGoGetDomains, ",")
-
-	// Clean-up noGoGetDomains to avoid needing to validate them when
+	// Parse noGoGetDomains to avoid needing to validate them when
 	// resolving static import paths
+	n.domains = parseCommaSeparatedList(conf.Get().NoGoGetDomains)
+}
+
+func parseCommaSeparatedList(list string) []string {
+	split := strings.Split(list, ",")
 	i := 0
-	for _, s := range n.domains {
+	for _, s := range split {
 		s = strings.TrimSpace(s)
 		if s != "" {
-			n.domains[i] = s
+			split[i] = s
 			i++
 		}
 	}
-	n.domains = n.domains[:i]
+	return split[:i]
 }
 
 // noGoGetDomains is a list of domains we do not attempt standard go vanity
@@ -93,6 +96,20 @@ func resolveStaticImportPath(importPath string) (*directory, error) {
 			vcs:         "git",
 			rev:         RuntimeVersion,
 		}, nil
+	}
+
+	// This allows users to set a list of domains that we should NEVER perform
+	// go get or git clone against. This is useful when e.g. a user has not
+	// correctly configured a monorepo and we are constantly hitting their
+	// production website to resolve import paths like "facebook.com/pkg/util"
+	// and skewing their own 404 metrics. This DOES mean these imports will be
+	// broken until they do correctly configure their monorepo (so we can
+	// identify its GOPATH), but it gives them a quick escape hatch that is
+	// better than "turn off the Sourcegraph server".
+	for _, domain := range conf.Get().BlacklistGoGet {
+		if strings.HasPrefix(importPath, domain) {
+			return nil, errors.New("import path in blacklistGoGet configuration")
+		}
 	}
 
 	// This allows a user to set a list of domains that are considered to be