+++ CHANGELOG.md
@@ -12,6 +12,9 @@ All notable changes to Sourcegraph Server and Data Center are documented in this
 * The new search filter `lang:` filters results by programming language (example: `foo lang:go` or `foo -lang:clojure`).
 * Dynamic filters: filters generated from your search results to help refine your results.
 * Search queries that consist only of `file:` now show files whose path matches the filters (instead of no results).
+* Sourcegraph now automatically detects basic `$GOPATH` configurations found in `.envrc` files in the root of repositories.
+* You can now configure the effective `$GOPATH`s of a repository by adding a `.sourcegraph/config.json` file to your repository with the contents `{"go": {"GOPATH": ["mygopath"]}}`.
+* A new `"blacklistGoGet": ["mydomain.org,myseconddomain.com"]` offers users a quick escape hatch in the event that Sourcegraph is making unwanted `go get` or `git clone` requests to their website due to incorrectly-configured monorepos. Most users will never use this option.
 
 ### Configuration changes